+++ src/YahooFantasyFootball/Services/IYahooApiService.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace YahooFantasyFootball.Services
+{
+    public interface IYahooApiService
+    {
+    }
+}
