+++ src/YahooFantasyFootball/Views/Home/Members.cshtml
@@ -1,13 +1,8 @@
 ﻿@model YahooFantasyFootball.ViewModels.MembersVM
 
 <h1>@ViewBag.Title</h1>
+<h2>@ViewBag.LeagueName</h2>
 <p>Welcome to my Fantasy Football League site</p>
 
-@Html.Partial("_SharedView")
 
-<form>
-    <label>Team Owner</label>
-    <textarea></textarea>
-    <lable>Placement</lable>
-    <textarea></textarea>
-</form>
+@Html.Partial("_SharedView")
